+++ src/components/detail-info/detail-info.test1.js
